+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on "Keep a Changelog".  This project adheres to Semantic Versioning.
 
 
+## [2.10.0] - 2023-01-17
+- Added new class called Daemon2.
+
+### Fixed
+- gen_class.TimeFormat:  Moved initial time hack from \_\_init\_\_ to create_adhoc_hack method.
+
+### Added
+- gen_libs.str_types: Returns basestring if pre-Python 3 or string if Python 3.
+- gen_class.Daemon2: Cleaner daemon code than the current Daemon class and works better in Python 3.
+
+### Changed
+- gen_class.TimeFormat.\_\_init\_\_: Added two new pre-defined data formats.
+
+
 ## [2.9.4] - 2022-09-06
 - Updated to work in Python 3 too.
 -     Note: gen_class.Yum class will only work in Python 2.7 at this time.